moved ntp_client into network class to be shared with other network functions
